+++ src/eca/llm_providers/ollama.clj
@@ -60,7 +60,7 @@
                    (fn [error-data]
                      (llm-util/log-response logger-tag rid "response-error" body)
                      (reset! response* error-data)))]
-    (llm-util/log-request logger-tag rid url body)
+    (llm-util/log-request logger-tag rid url body {})
     @(http/post
       url
       {:body (json/generate-string body)